+++ EnchantmentsEnhance-Core/src/main/kotlin/org/pixeltime/enchantmentsenhance/api/EnchantmentsEnhanceAPI.kt
@@ -40,8 +40,8 @@ class API : AbstractAPI {
         return ItemIcon.getOneStoneCountAsInt(playerName, stoneId)
     }
 
-    override fun addCustomEnchant(item: ItemStack, enchantment: String, level: Int) {
-        ItemManager.applyEnchantmentToItem(item, enchantment, level)
+    override fun addCustomEnchant(item: ItemStack, enchantment: String, level: Int) : Boolean{
+        return ItemManager.applyEnchantmentToItem(item, enchantment, level)
     }